+++ src/main/java/com/depot/ims/repositories/ItemsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ItemsRepository extends JpaRepository<Item,Integer> {
+public interface ItemsRepository extends JpaRepository<Item,Long> {
     List<Item> findAll();
 
     @Query("SELECT i FROM Item i WHERE i.itemName = ?1")
